refactor: integrate AccountRestrictionNotice into app/page.tsx
Fix module resolution for CurrencyFilter import.

#VERCEL_SKIP
