--- components/temporary-ban-alert/index.ts
@@ -1,1 +1,0 @@
-export { TemporaryBanAlert } from "./temporary-ban-alert"
